% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capsnet.R
\name{create_capsnet}
\alias{create_capsnet}
\title{Create capsnet model}
\usage{
create_capsnet(input_shape = c(28, 28, 1), n_class = 10L,
  num_routing = 3L)
}
\arguments{
\item{input_shape}{input_shape of images}

\item{n_class}{number of classes}

\item{num_routing}{number of routing units}
}
\description{
Create capsnet model
}
