% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capsnet.R
\name{primary_cap}
\alias{primary_cap}
\title{Apply Conv2D `n_channels` times and concatenate all capsules}
\usage{
primary_cap(inputs, dim_vector, n_channels, kernel_size, strides, padding)
}
\arguments{
\item{inputs}{4D tensor, shape=[None, width, height, channels]}

\item{dim_vector}{the dim of the output vector of capsule}

\item{n_channels}{the number of types of capsules}
}
\value{
output tensor, shape=[None, num_capsule, dim_vector]
}
\description{
Apply Conv2D `n_channels` times and concatenate all capsules
}
